AK: Use size_t for the length of strings
Using int was a mistake. This patch changes String, StringImpl,
StringView and StringBuilder to use size_t instead of int for lengths.
Obviously a lot of code needs to change as a result of this.

+++ AK/JsonParser.h
@@ -36,7 +36,7 @@ private:
     void consume_while(C);
 
     StringView m_input;
-    int m_index { 0 };
+    size_t m_index { 0 };
 
     String m_last_string_starting_with_character[256];
 };
