+++ Libraries/LibGUI/GWindow.h
@@ -175,7 +175,7 @@ private:
     bool m_show_titlebar { true };
     bool m_keybind_mode { false };
     String m_entered_keybind;
-    int m_max_keybind_length { 0 };
+    size_t m_max_keybind_length { 0 };
     HashMap<String, WeakPtr<GWidget>> m_keyboard_activation_targets;
     bool m_layout_pending { false };
 };
