+++ Libraries/LibPCIDB/Database.cpp
@@ -89,12 +89,12 @@ u8 parse_hex_digit(char digit)
 }
 
 template<typename T>
-T parse_hex(StringView str, int count)
+T parse_hex(StringView str, size_t count)
 {
     ASSERT(str.length() >= count);
 
     T res = 0;
-    for (int i = 0; i < count; i++)
+    for (size_t i = 0; i < count; i++)
         res = (res << 4) + parse_hex_digit(str[i]);
 
     return res;
